+++ client/src/Components/FileRecordsExportButton.tsx
@@ -9,9 +9,7 @@ import { teal } from '@material-ui/core/colors';
 import { saveAs } from 'file-saver';
 import { useSnackbar } from 'notistack';
 import React from 'react';
-import { Series } from '../gql/documents';
-import { useExportDataLazyQuery } from '../gql/queries';
-import { generateExcel } from '../utils/excel';
+import { useFileExportLazyQuery } from '../gql/queries';
 
 type Props = {
   setLoading: React.Dispatch<React.SetStateAction<boolean>>;
@@ -34,34 +32,47 @@ const useStyles = makeStyles((theme: Theme) =>
   })
 );
 
-export const ExportData = (props: Props) => {
+export const FileRecordsExportButton = (props: Props) => {
   const classes = useStyles();
   const { setLoading, loading: anyLoading } = props;
   const { enqueueSnackbar, closeSnackbar } = useSnackbar();
 
-  const [exportDataQuery] = useExportDataLazyQuery({
+  const [fileExportQuery] = useFileExportLazyQuery({
     fetchPolicy: 'no-cache',
     onCompleted: (data) => {
-      const file = generateExcel(data.allSeries as Series[])
-        .then((file) => saveAs(file, 'Exported Data.xlsx'))
-        .then(() => {
-          closeSnackbar('file-generation');
-          enqueueSnackbar(`Generated file.`, {
-            key: 'file-generation-success',
-            variant: 'success',
-          });
-          setLoading(false);
+      const content = data.files
+        ?.map((file) => file?.path)
+        .filter(Boolean)
+        .reduce((previous, current) => {
+          return previous + `"${current}",\r\n`;
+        }, '');
+      closeSnackbar('file-generation');
+      if (content) {
+        const file = new File([content], 'fileList.csv', {
+          type: 'text/csv,charset=utf-8',
         });
+        enqueueSnackbar(`Generated file.`, {
+          key: 'file-generation-success',
+          variant: 'success',
+        });
+        saveAs(file);
+      } else {
+        enqueueSnackbar(`Something went wrong. Please try again.`, {
+          key: 'file-generation',
+          variant: 'warning',
+        });
+      }
+      setLoading(false);
     },
   });
 
-  const handleClick = async () => {
+  const handleClick = () => {
     setLoading(true);
     enqueueSnackbar(`Generating file... Please wait...`, {
       key: 'file-generation',
       autoHideDuration: null,
     });
-    exportDataQuery();
+    fileExportQuery();
   };
 
   return (
@@ -71,7 +82,7 @@ export const ExportData = (props: Props) => {
         disabled={anyLoading}
         variant="contained"
       >
-        Export Data
+        Export File List CSV
       </Button>
       {anyLoading && (
         <CircularProgress size={24} className={classes.buttonProgress} />
